requirements.txt
flask
flask-migrate
flask-sqlalchemy
sqlalchemy
python-dotenv
psycopg2
jwt
pyjwt